type registerIconSet API
