+++ stores/icons/actions.ts
@@ -1,27 +1,21 @@
+import { IconSet, ActionTypes } from "./types";
+
 /**
  * Returns an action object used in signalling that new block styles have been added.
- *
- * @param {object} iconSet icon set.
- *
- * @returns {object} Action object.
  */
-export function registerIconSet(iconSet) {
+export function registerIconSet(iconSet: IconSet) {
 	return {
-		type: 'REGISTER_ICON_SET',
+		type: ActionTypes.REGISTER_ICON_SET,
 		iconSet,
 	};
 }
 
 /**
  * Returns an action object used in signalling that block styles have been removed.
- *
- * @param {string} name  Icon Set name.
- *
- * @returns {object} Action object.
  */
-export function removeIconSet(name) {
+export function removeIconSet(name: string) {
 	return {
-		type: 'REMOVE_ICON_SET',
+		type: ActionTypes.REMOVE_ICON_SET,
 		name,
 	};
 }
